methods implemented in UserServiceImpl class, some method created in userJPA Interface

+++ Rest-Api/src/main/java/com/wedlock/jpa/FeedbackJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Feedback;
 
-public interface FeedbackJPA extends JpaRepository<Feedback, Integer>{
+public interface FeedbackJPA extends JpaRepository<Feedback, Long>{
 
 }
