+++ Rest-Api/src/main/java/com/wedlock/jpa/MatchJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Match;
 
-public interface MatchJPA extends JpaRepository<Match, Integer>{
+public interface MatchJPA extends JpaRepository<Match, Long>{
 
 }
