+++ Rest-Api/src/main/java/com/wedlock/jpa/MessageJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Message;
 
-public interface MessageJPA extends JpaRepository<Message, Integer>{
+public interface MessageJPA extends JpaRepository<Message, Long>{
 
 }
