+++ Rest-Api/src/main/java/com/wedlock/jpa/PaymentJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Payment;
 
-public interface PaymentJPA extends JpaRepository<Payment, Integer>{
+public interface PaymentJPA extends JpaRepository<Payment, Long>{
 
 }
