+++ Rest-Api/src/main/java/com/wedlock/jpa/PhotoJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Photo;
 
-public interface PhotoJPA extends JpaRepository<Photo, Integer>{
+public interface PhotoJPA extends JpaRepository<Photo, Long>{
 
 }
