+++ Rest-Api/src/main/java/com/wedlock/jpa/ReportJPA.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wedlock.entities.Report;
 
-public interface ReportJPA extends JpaRepository<Report, Integer>{
+public interface ReportJPA extends JpaRepository<Report, Long>{
 
 }
